requirements.txt
#Langchain Core

langchain
langchain-core

# Open AI Integration 
langchain_openai
openai

#Anthropic integration
langchain-anthropic

#Google services
langchain-google-genai
google-generativeai

#Hugging Face Integration
langchain-huggingface
transformers
huggingface-hub

#Environment varible management
python-dotenv 

#Machine Learning
numpy 
scikit-learn

#For prompt UI 
streamlit




